Use std::size_t for RobotPool::getAvailableCount

+++ OPHD/States/MapViewStateDraw.cpp
@@ -67,7 +67,7 @@ void MapViewState::drawRobotInfo() const
 		std::tuple{minerImageRect, mRobotPool.getAvailableCount(Robot::Type::Miner), mRobotPool.miners().size()},
 		std::tuple{dozerImageRect, mRobotPool.getAvailableCount(Robot::Type::Dozer), mRobotPool.dozers().size()},
 		std::tuple{diggerImageRect, mRobotPool.getAvailableCount(Robot::Type::Digger), mRobotPool.diggers().size()},
-		std::tuple{robotSummaryImageRect, static_cast<int>(mRobotPool.currentControlCount()), mRobotPool.robotControlMax()},
+		std::tuple{robotSummaryImageRect, mRobotPool.currentControlCount(), mRobotPool.robotControlMax()},
 	};
 
 	for (const auto& [imageRect, parts, total] : icons)
